fixup

+++ tests/server/mqttd.c
@@ -155,7 +155,7 @@ static void logprotocol(mqttdir dir,
   }
   fprintf(output, "%s %s %x %s\n",
           dir == FROM_CLIENT ? "client" : "server",
-          prefix, remlen, data);
+          prefix, (unsigned int)remlen, data);
 }
 
 /* return 0 on success */
